# perf: allow pre-allocate of size when new map