There is a simple result from the bench test:
|entries|withInitialedSize|withoutInitialedSize|rate|
|-------|------|----|---|
| 1e3 | 41153 ns/op  | 72909 ns/op | 56.44% |
|1e5 | 7523395 ns/op | 9731575 ns/op | 77.31% |
|1e7 | 2005930041 ns/op | 2555268041 ns/op | 78.50% |

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/elliotchance/orderedmap/48)
<!-- Reviewable:end -->